# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |
    Share the moments you love with people you love!
  version: 2.0.2
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    uid:
      name: uid
      description: User identifier given by server at the moment of the sign up
      in: path
      schema:
        $ref: "#/components/schemas/user/properties/id"
      required: true
...
  schemas:
    photo:
      type: object
      description: Photo Object
      properties:
        photo_data:
          type: string
          format: binary
          description: photo binary data
          minLength: 0
          maxLength: 100
        username_owner:
          $ref: "#/components/schemas/user/properties/name"
    profile:
      type: object
      description: User profile infos
      properties:
        username:
          $ref: "#/components/schemas/user/properties/name"
        id:
          $ref: "#/components/schemas/user/properties/id"
        following:
          type: integer
          description: following amount
          example: 12
        followers:
          type: integer
          description: followers amount
          example: 12
        posts_amount:
          type: integer
          description: posts amount
          example: 12
...
  responses:
    ServerDown:
      description: "Server is down"
      content:
        application/json:
          schema:
            type: string
            description: Log
            example: "Server Down"
    BadRequest:
      description: "Bad request"
      content:
        application/json:
          schema:
            type: string
            description: Log
            example: "Bad request"
    Success:
      description: "Success"
      content:
        application/json:
          schema:
            type: string
            description: Log
            example: "Success"
    DeleteSuccess:
      description: "Resource deleted successfully"
      content:
        application/json:
          schema:
            type: string
            description: Log
            example: "Deleted"
paths:
...
  /user/{uid}/username:
    # setMyUserName
    parameters:
     - $ref: "#/components/parameters/uid"
    put:
      tags: ["User"]
      summary: "Change my username"
      description: If a user want to change his username he can patch it
      operationId: setMyUserName
      requestBody:
        description: New Username
        content:
          application/json:
            schema:
              type: object
              description: request object
              properties:
                newUsername:
                  $ref: "#/components/schemas/user/properties/name"
        required: true
      responses:
        '200':
          $ref : "#/components/responses/Success"
        '400':
          $ref : "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerDown"
      security:
        - bearerAuth: [] 
  /user/{uid}/following:
    # followUser
    parameters:
      - $ref: "#/components/parameters/uid"
    post:
      tags: ["User"]
      operationId: followUser
      summary: "Follow a user"
      description: A user can change his username specifying it in the requestBody
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: request object
              properties:
                username:
                  $ref: "#/components/schemas/user/properties/name"
        required: true
      responses:
        '201':
          description: Successfully added in following collection
          content:
            application/json:
              schema:
                type: string
                description: Log
                example: "Success"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerDown"
      security:
        - bearerAuth: [] 
  /user/{uid}/stream:
    #getMyStream
    parameters:
      - $ref: "#/components/parameters/uid"
    get:
      summary: Get user stream
      description: Get user stream
      tags: ["Photo"]
      operationId: getMyStream
      responses:
        '200':
          description: Return stream of photo from following
          content:
            application/json:
              schema:
                type: object
                description: response object
                properties:
                  stream:
                    type: array
                    description: Array of photo
                    items:
                      $ref: "#/components/schemas/photo"
                    minItems: 0
                    maxItems: 10
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerDown"
      security:
        - bearerAuth: [] 
  /user/{uid}/following/{following_uid}:
    # unfollowUser
    parameters:
      - $ref: "#/components/parameters/uid"
      - name: following_uid
        description: User identifier
        in: path
        schema:
          $ref: "#/components/schemas/user/properties/id"
        required: true
    delete:
      tags: ["User"]
      summary: "Unfollow a user"
      description: Delete a follower from your following users
      operationId: unfollowUser
      responses:
        '200':
          $ref: "#/components/responses/DeleteSuccess"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerDown"
      security:
          - bearerAuth: [] 
  /user/{uid}/ban:
    #banUser
    parameters:
      - $ref: "#/components/parameters/uid"
    post:
      summary: Ban user
      description: Put a user in your own banned list
      operationId: banUser
      tags: ["User"]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: request object
              properties:
                username:
                  $ref: "#/components/schemas/user/properties/name"
        required: true
      responses:
        '201':
          description: User identifier has been banned
          content:
            application/json:
              schema:
                type: object
                description: request Object
                properties:
                  message:
                    type: string
                    description: Log
                    example: "Success"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerDown"
      security:
        - bearerAuth: [] 
  /user/{uid}/ban/{banned_uid}:
    #unbanUser
    parameters:
      - $ref: "#/components/parameters/uid"
      - name: banned_uid
        description: User identifier
        in: path
        schema:
          $ref: "#/components/schemas/user/properties/id"
        required: true
    delete:
      summary: Unban user
      description: Remove a user from your banned list
      tags: ["User"]
      operationId: unbanUser
      responses:
        '200':
          $ref: "#/components/responses/DeleteSuccess"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerDown"
      security:
        - bearerAuth: [] 
  /user/{uid}:
    #getUserProfile
    parameters:
      - $ref: "#/components/parameters/uid"
    get:
      summary: Get user profile info
      description: Get user profile info by user id
      operationId: getUserProfile
      tags: ["User"]
      responses:
        '200':
          description: Return user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/profile"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerDown"
      security:
        - bearerAuth: [] 
...